P4PU-1050 updated query
